% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statanalysis_functions.R
\name{filter.counts}
\alias{filter.counts}
\title{Performing filtering}
\usage{
filter.counts(countTable)
}
\arguments{
\item{countTable}{Variable storing the count data table}
}
\value{
Count table with low filtered values removed
}
\description{
Filters by rowmeans(log2(cpm())+1)
}
\examples{
filterdata <- filtering(countTable);
}
