% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statanalysis_functions.R
\name{linmodel}
\alias{linmodel}
\title{Linear Models}
\usage{
linmodel(countTable, sampleTable)
}
\arguments{
\item{countTable, sampleTable}{Count and sample data variable to dge normalize and make design matrix andf linear modelling}
}
\value{
Data files and png files for enrichment analysis visual representation
}
\description{
Makes design matrix and Fits linear models and performs stat testing
}
\examples{
linearmodel <- linmodel(countTable,sampleTable);
}
