% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statanalysis_functions.R
\name{read.counts}
\alias{read.counts}
\title{Reading counts}
\usage{
read.counts(counts)
}
\arguments{
\item{counts}{The name/path of the file in quotes}
}
\value{
Count data table
}
\description{
Count data file is read in
}
\examples{
count_data <- countData("E-MTAB-2523.counts.txt");

}
