% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statanalysis_functions.R
\name{read.sample}
\alias{read.sample}
\title{Sample data reading}
\usage{
read.sample(sample)
}
\arguments{
\item{sample}{The name/path to the sample data file in quotes}
}
\value{
The sample data and the factors
}
\description{
Reads in the data of the sample table and creates factors for the information
}
\examples{
sample_data <- sampleData("E-MTAB-2523_sample table.txt");
}
